GP2-3287:  Add 'related opportunities' to the data available for a Region page
This changeset expands (but does not otherwise amend) the API representation of a Region page, so that it includes up to three serializes Investment Opportunities in its response. (ie, the Opps are Atlas ones, not the old CapInvest type)

To be selected, an InvestmentOpportunity must have the current region as one of their related regions. They are then sorted by priority_weighting, if set, and then use their PK as a secondary sorting key. This means that highly weighted ones appear first, then any ties are broken by using the most recently created (ie highest PK) opp first.

In total up to three opps are included, but 0..3 is possible.